add z layer

+++ src/ascii.cpp
@@ -34,10 +34,10 @@ struct ascii_ {
 
 static struct ascii_ ascii;
 
-int16_t TERM_WIDTH;
-int16_t TERM_HEIGHT;
-int16_t ascii_mouse_x;
-int16_t ascii_mouse_y;
+int TERM_WIDTH;
+int TERM_HEIGHT;
+int ascii_mouse_x;
+int ascii_mouse_y;
 
 std::array< std::array< AsciiCell, TERM_HEIGHT_MAX >, TERM_WIDTH_MAX > *cells;
